Move validation tests into own file
Also add regression test for #1211.
